Refactoring and more docs
* Some much needed refactoring to both the add-on folder and the tests
* Update docs listed in #40
